# Use `rb_enc_get()` to retrieve encoding type