@Watson1978 2021-08-08
I found the same code in another place. https://github.com/ohler55/oj/blob/4fdd25d408336ab05ae3d16bb81fb790850d9074/ext/oj/parse.c#L966-L971